Rid of LinkLike and LinkWalker; Simplify!
